LayoutTests/imported/w3c/web-platform-tests/content-security-policy/script-src-attr-elem/script-src-elem-blocked-src-allowed-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'script-src-elem'.


Harness Error (TIMEOUT), message = null

NOTRUN Should fire a spv event
